# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EditalInput {
  /**
     * Raw edital text to be simplified
     * @minLength 10
     */
  text: string;
}

export interface ExtractUrlInput {
  /** Public URL of the edital to extract text from */
  url: string;
}

export interface ExtractedText {
  /** Extracted text content */
  text: string;
  /** Page title if available */
  title: string;
}

export interface SimplifiedEdital {
  resumo: string;
  objetivo: string;
  quemPodeParticipar: string;
  prazoInscricao: string;
  ondeSeInscrever: string;
  principaisRequisitos: string;
  linguagemSimples: string;
}

export interface SaveEditalInput {
  title: string;
  originalText: string;
  resumo: string;
  objetivo: string;
  quemPodeParticipar: string;
  prazoInscricao: string;
  ondeSeInscrever: string;
  principaisRequisitos: string;
  linguagemSimples: string;
}

export interface SavedEdital {
  id: number;
  title: string;
  originalText: string;
  resumo: string;
  objetivo: string;
  quemPodeParticipar: string;
  prazoInscricao: string;
  ondeSeInscrever: string;
  principaisRequisitos: string;
  linguagemSimples: string;
  createdAt: string;
}

export type SaveAgentResultInputResultJson = { [key: string]: unknown };

export interface SaveAgentResultInput {
  agentId: string;
  title: string;
  originalText: string;
  resultJson: SaveAgentResultInputResultJson;
}

export type AgentResultRecordResultJson = { [key: string]: unknown };

export interface AgentResultRecord {
  id: number;
  agentId: string;
  title: string;
  originalText: string;
  resultJson: AgentResultRecordResultJson;
  createdAt: string;
}

export interface AgentUserProfile {
  escolaridade?: string;
  atuacao?: string;
  municipio?: string;
  rendaFamiliar?: string;
}

export type AgentAnalyzeInputAgentId = typeof AgentAnalyzeInputAgentId[keyof typeof AgentAnalyzeInputAgentId];


export const AgentAnalyzeInputAgentId = {
  simples: 'simples',
  analista: 'analista',
  estrategica: 'estrategica',
  acompanhamento: 'acompanhamento',
  documentacao: 'documentacao',
  elegibilidade: 'elegibilidade',
} as const;

export interface AgentAnalyzeInput {
  agentId: AgentAnalyzeInputAgentId;
  /**
     * Raw edital text to be analyzed
     * @minLength 10
     */
  text: string;
  profile?: AgentUserProfile;
}

export interface AgentAnalyzeResponse {
  type: string;
  [key: string]: unknown;
 }

export type EditalStatsByAgentItem = {
  agentId: string;
  count: number;
};

export interface EditalStats {
  total: number;
  byAgent: EditalStatsByAgentItem[];
  recentAnalyses: AgentResultRecord[];
}

export interface ErrorResponse {
  error: string;
}