# .migration-backup/lib/api-zod/src/generated/types/agentUserProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentUserProfile {
  escolaridade?: string;
  atuacao?: string;
  municipio?: string;
  rendaFamiliar?: string;
}