# .migration-backup/lib/api-zod/src/generated/types/saveEditalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveEditalInput {
  title: string;
  originalText: string;
  resumo: string;
  objetivo: string;
  quemPodeParticipar: string;
  prazoInscricao: string;
  ondeSeInscrever: string;
  principaisRequisitos: string;
  linguagemSimples: string;
}